nfc_linux: split fido_nfc_[rt]x into nfc.c
these will be reused by upcoming pcsc support
